requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
fastapi==0.115.4
h11==0.14.0
idna==3.10
jinja2==3.1.4
markupsafe==3.0.2
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pyjwt==2.9.0
python-multipart==0.0.16
ruff==0.7.1
sniffio==1.3.1
starlette==0.41.2
typing-extensions==4.12.2
uvicorn==0.32.0
